move task_manager to x-pack/plugins
