+++ src/server/config/schema.js
@@ -244,30 +244,9 @@ export default () => Joi.object({
     }),
     profile: Joi.boolean().default(false)
   }).default(),
-
   status: Joi.object({
     allowAnonymous: Joi.boolean().default(false)
   }).default(),
-
-  task_manager: Joi.object({
-    max_attempts: Joi.number()
-      .description('The maximum number of times a task will be attempted before being abandoned as failed')
-      .default(3),
-    poll_interval: Joi.number()
-      .description('How often, in milliseconds, the task manager will look for more work.')
-      .default(3000),
-    index: Joi.string()
-      .description('The name of the index used to store task information.')
-      .default('.kibana_task_manager'),
-    max_workers: Joi.number()
-      .description('The maximum number of tasks that this Kibana instance will run simultaneously.')
-      .default(10),
-    override_num_workers: Joi.object()
-      .pattern(/.*/, Joi.number().greater(0))
-      .description('Customize the number of workers occupied by specific tasks (e.g. override_num_workers.reporting: 2)')
-      .default({})
-  }).default(),
-
   map: Joi.object({
     includeElasticMapsService: Joi.boolean().default(true),
     tilemap: tilemapSchema,
