+++ src/server/kbn_server.js
@@ -38,7 +38,6 @@ import * as Plugins from './plugins';
 import { indexPatternsMixin } from './index_patterns';
 import { savedObjectsMixin } from './saved_objects';
 import { sampleDataMixin } from './sample_data';
-import { taskManagerMixin } from './task_manager';
 import { KibanaMigrator } from './saved_objects/migrations';
 import { urlShorteningMixin } from './url_shortening';
 import { serverExtensionsMixin } from './server_extensions';
@@ -94,9 +93,6 @@ export default class KbnServer {
 
       // setup routes for installing/uninstalling sample data sets
       sampleDataMixin,
-
-      // task manager service
-      taskManagerMixin,
 
       // setup routes for short urls
       urlShorteningMixin,
