+++ x-pack/plugins/task_manager/lib/extract_task_definitions.test.ts
@@ -1,20 +1,7 @@
 /*
- * Licensed to Elasticsearch B.V. under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch B.V. licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
  */
 
 import { get } from 'lodash';
@@ -58,7 +45,7 @@ describe('extractTaskDefinitions', () => {
     const maxWorkers = 10;
     const overrideNumWorkers = {};
     const taskDefinitions = getMockTaskDefinitions({ numTasks: 3 });
-    const result = extractTaskDefinitions(maxWorkers, taskDefinitions, overrideNumWorkers);
+    const result = extractTaskDefinitions(taskDefinitions, maxWorkers, overrideNumWorkers);
 
     expect(result).toMatchInlineSnapshot(`
 Object {
@@ -94,7 +81,7 @@ Object {
     const maxWorkers = 2;
     const overrideNumWorkers = {};
     const taskDefinitions = getMockTaskDefinitions({ numTasks: 2, numWorkers: 5 });
-    const result = extractTaskDefinitions(maxWorkers, taskDefinitions, overrideNumWorkers);
+    const result = extractTaskDefinitions(taskDefinitions, maxWorkers, overrideNumWorkers);
 
     expect(result).toMatchInlineSnapshot(`
 Object {
@@ -125,7 +112,7 @@ Object {
     };
     const maxWorkers = 5;
     const taskDefinitions = getMockTaskDefinitions({ numTasks: 3 });
-    const result = extractTaskDefinitions(maxWorkers, taskDefinitions, overrideNumWorkers);
+    const result = extractTaskDefinitions(taskDefinitions, maxWorkers, overrideNumWorkers);
 
     expect(result).toMatchInlineSnapshot(`
 Object {
@@ -177,7 +164,7 @@ Object {
         },
       };
 
-      return extractTaskDefinitions(maxWorkers, taskDefinitions, overrideNumWorkers);
+      return extractTaskDefinitions(taskDefinitions, maxWorkers, overrideNumWorkers);
     };
 
     expect(runExtract).toThrowError();
