+++ x-pack/plugins/task_manager/lib/extract_task_definitions.ts
@@ -1,20 +1,7 @@
 /*
- * Licensed to Elasticsearch B.V. under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch B.V. licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
  */
 
 import Joi from 'joi';
@@ -36,8 +23,8 @@ import {
  * @param overrideNumWorkers - The kibana.yml overrides numWorkers per task type.
  */
 export function extractTaskDefinitions(
-  maxWorkers: number,
   taskDefinitions: TaskDictionary<TaskDefinition> = {},
+  maxWorkers: number,
   overrideNumWorkers: { [taskType: string]: number }
 ): TaskDictionary<SanitizedTaskDefinition> {
   return Object.keys(taskDefinitions).reduce(
